test: document stub function checks
