+++ docs/master_checklist.md
@@ -4,7 +4,7 @@ Consult [`design_overview.md`](design_overview.md) for overall requirements.
 Execute each step in order, marking completion after satisfying the [Definition of Done](definition_of_done.md).
 
 1. [x] [Step 0: Project Setup](steps/00_project_setup.md)
-2. [ ] [Step 1: Core World Generation](steps/01_core_world.md)
+2. [ ] [Step 1: Core World Generation](steps/01_core_world.md) – in progress
 3. [ ] [Step 2: Transport Network](steps/02_transport.md)
 4. [ ] [Step 3: Land Use and Settlements](steps/03_growth.md)
 5. [ ] [Step 4: Industry Sites](steps/04_industries.md)
