+++ src/export/gif.ts
@@ -1,4 +1,4 @@
-import { Sim } from './types';
+import { Sim } from '../types';
 
 export function captureIfDue(sim: Sim, ctx: CanvasRenderingContext2D): void {
   throw new Error('captureIfDue not implemented');
