--- src/growth.ts
@@ -1,13 +1,0 @@
-import { Sim } from './types';
-
-export function updateLandUse(sim: Sim): void {
-  throw new Error('updateLandUse not implemented');
-}
-
-export function updateSettlements(sim: Sim): void {
-  throw new Error('updateSettlements not implemented');
-}
-
-export function updateIndustries(sim: Sim): void {
-  throw new Error('updateIndustries not implemented');
-}
