+++ src/industries/site_select.ts
@@ -1,0 +1,5 @@
+import { Sim } from '../types';
+
+export function updateIndustries(sim: Sim): void {
+  throw new Error('updateIndustries not implemented');
+}
