+++ src/landuse/update.ts
@@ -1,0 +1,5 @@
+import { Sim } from '../types';
+
+export function updateLandUse(sim: Sim): void {
+  throw new Error('updateLandUse not implemented');
+}
