+++ src/render/index.ts
@@ -1,4 +1,4 @@
-import { Sim } from './types';
+import { Sim } from '../types';
 
 export function renderFrame(sim: Sim, ctx: WebGL2RenderingContext): void {
   throw new Error('renderFrame not implemented');
