+++ src/sim/upgrades.ts
@@ -1,0 +1,5 @@
+import { Sim } from '../types';
+
+export function applyUpgrades(sim: Sim): void {
+  throw new Error('applyUpgrades not implemented');
+}
