+++ src/society/settlements.ts
@@ -1,0 +1,5 @@
+import { Sim } from '../types';
+
+export function updateSettlements(sim: Sim): void {
+  throw new Error('updateSettlements not implemented');
+}
